requirements.txt
# BINAUTOGO - Python Dependencies
# Совместимо с Python 3.7 - 3.11

# ============================================
# ОСНОВНЫЕ БИБЛИОТЕКИ
# ============================================
requests>=2.31.0
websocket-client>=1.6.4
pandas>=2.0.0
numpy>=1.24.0

# ============================================
# КРИПТОВАЛЮТНЫЕ БИРЖИ
# ============================================
python-binance>=1.0.19
ccxt>=4.1.77

# ============================================
# ПЛАНИРОВАНИЕ
# ============================================
schedule>=1.2.0
python-dateutil>=2.8.2

# ============================================
# КОНФИГУРАЦИЯ
# ============================================
python-dotenv>=1.0.0
jsonschema>=4.20.0

# ============================================
# TELEGRAM BOT
# ============================================
python-telegram-bot>=20.7

# ============================================
# WEB DASHBOARD (Cyberpunk Style!)
# ============================================
streamlit>=1.29.0
plotly>=5.18.0
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0

# ============================================
# MACHINE LEARNING
# ============================================
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0

# ============================================
# SENTIMENT ANALYSIS
# ============================================
tweepy>=4.14.0
praw>=7.7.1
textblob>=0.17.1
vaderSentiment>=3.3.2

# ============================================
# АСИНХРОННОСТЬ
# ============================================
aiohttp>=3.9.1
asyncio>=3.4.3

# ============================================
# БЕЗОПАСНОСТЬ
# ============================================
cryptography>=41.0.7

# ============================================
# ПРОИЗВОДИТЕЛЬНОСТЬ
# ============================================
psutil>=5.9.6
diskcache>=5.6.3

# ============================================
# ПРИМЕЧАНИЯ
# ============================================
# Установка:
#   pip install -r requirements.txt
#
# Для Python 3.7:
#   pip install -r requirements.txt --upgrade
#
# Для GPU поддержки (опционально):
#   pip install tensorflow-gpu  # Если нужен TensorFlow
